Game Infomd
The Secret Number Game is a project I develop using HTML, CSS and JavaScritp in order to show my skills with programming logic.

The goal is to find what is the secret number. Once you done, the game will display a message and show how many attempts were made to find the secret number.

Account creation is not required.

Hope you enjoy it!